Fix overflow bug on append
Now use size_t as the type for row_counts.
Also adds a todo to make shape_t unsigned and to use uint64_t for all
row counts (not done now as it is a non-trivial change).
Test from previous commit now passes.

+++ cpp/arcticdb/pipeline/index_utils.cpp
@@ -80,7 +80,7 @@ std::pair<index::IndexSegmentReader, std::vector<SliceAndKey>> read_index_to_vec
 }
 
 TimeseriesDescriptor get_merged_tsd(
-        int row_count,
+        size_t row_count,
         bool dynamic_schema,
         const TimeseriesDescriptor& existing_tsd,
         const std::shared_ptr<pipelines::InputTensorFrame>& new_frame) {
